requirements.txt
memory-profiler==0.61.0
mysql-connector-python==9.0.0
psutil==6.0.0
